pyproject.toml
[project]
name = "finance-insight-service"
version = "0.1.0"
description = "Finance Insight Service - research agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "crewai[tools]>=0.83.0",
  "crewai-tools>=0.28.0",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "python-dotenv>=1.0.1",
  "serpapi>=0.1.5",
]

[project.scripts]
research_agent = "finance_insight_service.main:main"
quant_agent = "finance_insight_service.quant_main:main"
finance_insight = "finance_insight_service.full_main:main"

[tool.uv]
package = true
